Add files via upload
Issue: firerate capped at a cycle time of .06 seconds per cycle, not nearly fast enough. When shooting a wall, (ignore the case where it hits a player), most of the magazine shoots with a .06 second long cycle however randomly ~4-8 shots will shoot at the correct speed in a row, and then it reverts back to a shitty cycle time. Good luck chief.